requirements.txt
streamlit
strands-agents